fix unit check
A systemd unit has several state fields.
1. `LoadState` tells where the configuration file of this unit has been loaded.
   We consider a unit unhealthy if LoadState != "loaded"

2. `ActiveState` tells whether the unit is `active` or not. Possbile states:
   active, reloading, inactive, failed, activating, deactivating. We consider
   `active`, `inactive` and `activating*` to be healthy active states.

3. `SubState` a more detailed state description. Should be `running` for a
   healthy state. And should be `dead` for one shot.

*This change adds slightly more checks for `activating` state listed in 2.
When the service is in `activating` state it may indicate that the service is
unable to start (become active). The reason for failure might be a failing
ExecStartPre script. To make sure we catch that we are using the following
logic.

A unit is considered to be unhealthy if:
 - `ActiveState` is `activating` and `SubState` is `auto-restart` and
   (`ActiveEnterTimestampMonotonic` is 0 or \
      `InactiveEnterTimestampMonotonic`  > `ActiveEnterTimestampMonotonic`)
   that means that a unit either can't start at all for the first time or
   unit was able to start before but now it can't

for ref: https://www.freedesktop.org/wiki/Software/systemd/dbus/